Added: Deprecation warning for Ruby < 2.0.
