+++ spec/ddtrace/configuration_spec.rb
@@ -1,131 +1,43 @@
 require 'spec_helper'
 
-require 'ddtrace'
+require 'ddtrace/patcher'
+require 'ddtrace/configuration'
 
-RSpec.describe Datadog::Configuration::Settings do
-  let(:configuration) { described_class.new(registry: registry) }
-  let(:registry) { Datadog::Contrib::Registry.new }
+RSpec.describe Datadog::Configuration do
+  context 'when extended by a class' do
+    subject(:test_class) { stub_const('TestClass', Class.new { extend Datadog::Configuration }) }
 
-  describe '#use' do
-    subject(:result) { configuration.use(name, options) }
-    let(:name) { :example }
-    let(:integration) { double('integration') }
-    let(:options) { {} }
+    describe '#configure' do
+      subject(:configure) { test_class.configure }
 
-    before(:each) do
-      registry.add(name, integration)
-    end
-
-    context 'for a generic integration' do
-      before(:each) do
-        expect(integration).to receive(:configure).with(:default, options).and_return([])
-        expect(integration).to receive(:patch).and_return(true)
-      end
-
-      it { expect { result }.to_not raise_error }
-    end
-
-    context 'for an integration that includes Datadog::Contrib::Integration' do
-      let(:patcher_module) do
-        stub_const('Patcher', Module.new do
-          include Datadog::Contrib::Patcher
-
-          def self.patch
-            true
-          end
-        end)
-      end
-
-      let(:integration_class) do
-        patcher_module
-
-        Class.new do
-          include Datadog::Contrib::Integration
-
-          def self.version
-            Gem::Version.new('0.1')
-          end
-
-          def patcher
-            Patcher
+      context 'deprecation warning for Ruby 1.9' do
+        before do
+          # Reset deprecation warning
+          if Datadog::Patcher.instance_variable_defined?(:@done_once) \
+             && !Datadog::Patcher.instance_variable_get(:@done_once).nil?
+            Datadog::Patcher.instance_variable_get(:@done_once).delete(:ruby_19_deprecation_warning)
           end
         end
-      end
 
-      let(:integration) do
-        integration_class.new(name)
-      end
+        if Gem::Version.new(RUBY_VERSION) < Gem::Version.new('2.0')
+          context 'is raised' do
+            it do
+              expect(Datadog::Tracer.log).to receive(:warn)
+                .with(described_class::RUBY_19_DEPRECATION_WARNING)
 
-      context 'which is provided only a name' do
-        it do
-          expect(integration).to receive(:configure).with(:default, {})
-          configuration.use(name)
+              configure
+            end
+          end
+        else
+          context 'is not raised' do
+            it do
+              expect(Datadog::Tracer.log).to_not receive(:warn)
+                .with(described_class::RUBY_19_DEPRECATION_WARNING)
+
+              configure
+            end
+          end
         end
-      end
-
-      context 'which is provided a block' do
-        it do
-          expect(integration).to receive(:configure).with(:default, {}).and_call_original
-          expect { |b| configuration.use(name, options, &b) }.to yield_with_args(
-            a_kind_of(Datadog::Contrib::Configuration::Settings)
-          )
-        end
-      end
-    end
-  end
-
-  describe '#tracer' do
-    let(:tracer) { Datadog::Tracer.new }
-    let(:debug_state) { tracer.class.debug_logging }
-    let(:custom_log) { Logger.new(STDOUT) }
-
-    context 'given some settings' do
-      before(:each) do
-        @original_log = tracer.class.log
-
-        configuration.tracer(
-          enabled: false,
-          debug: !debug_state,
-          log: custom_log,
-          hostname: 'tracer.host.com',
-          port: 1234,
-          env: :config_test,
-          tags: { foo: :bar },
-          instance: tracer
-        )
-      end
-
-      after(:each) do
-        tracer.class.debug_logging = debug_state
-        tracer.class.log = @original_log
-      end
-
-      it 'applies settings correctly' do
-        expect(tracer.enabled).to be false
-        expect(debug_state).to be false
-        expect(Datadog::Tracer.log).to eq(custom_log)
-        expect(tracer.writer.transport.current_api.adapter.hostname).to eq('tracer.host.com')
-        expect(tracer.writer.transport.current_api.adapter.port).to eq(1234)
-        expect(tracer.tags[:env]).to eq(:config_test)
-        expect(tracer.tags[:foo]).to eq(:bar)
-      end
-    end
-
-    it 'acts on the default tracer' do
-      previous_state = Datadog.tracer.enabled
-      configuration.tracer(enabled: !previous_state)
-      expect(Datadog.tracer.enabled).to eq(!previous_state)
-      configuration.tracer(enabled: previous_state)
-      expect(Datadog.tracer.enabled).to eq(previous_state)
-    end
-  end
-
-  describe '#[]' do
-    context 'when the integration doesn\'t exist' do
-      it do
-        expect { configuration[:foobar] }.to raise_error(
-          Datadog::Contrib::Extensions::Configuration::Settings::InvalidIntegrationError
-        )
       end
     end
   end
